LayoutTests/inspector/canvas/create-context-webgl-expected.txt

Test that CanvasManager tracks creation and destruction of WebGL canvases.


== Running test suite: Canvas.CreateContextWebGL
-- Running test case: Canvas.CreateContextWebGL.NoCanvases
PASS: CanvasManager should have no canvases.

-- Running test case: Canvas.CreateContextWebGL.Attached
PASS: Canvas context should be WebGL.
PASS: Removed canvas has expected ID.

-- Running test case: Canvas.CreateContextWebGL.Detached
PASS: Canvas context should be WebGL.
PASS: Removed canvas has expected ID.

-- Running test case: Canvas.CreateContextWebGL.CSSCanvas
Create CSS canvas from -webkit-canvas(css-canvas).
PASS: Canvas context should be WebGL.
PASS: Canvas name should equal the identifier passed to -webkit-canvas.
